testing.benchmarks: separate fixtures and config
