+++ pyproject.toml
@@ -87,7 +87,11 @@ s3 = ["dvc-s3==2.22.0"]
 ssh = ["dvc-ssh>=2.22.1,<3"]
 ssh_gssapi = ["dvc-ssh[gssapi]>=2.22.1,<3"]
 terraform = ["tpi[ssh]>=2.1"]
-testing = ["pytest-test-utils"]
+testing = [
+  "pytest-test-utils",
+  "pytest-benchmark[histogram]",
+  "pytest-virtualenv",
+]
 tests = [
     "beautifulsoup4>=4.4",
     "dvc-ssh",
@@ -114,6 +118,8 @@ Source = "https://github.com/iterative/dvc"
 dvc = "dvc.cli:main"
 [project.entry-points."fsspec.specs"]
 dvc = "dvc.api:DVCFileSystem"
+[project.entry-points."pytest11"]
+dvc-testing = "dvc.testing"
 
 [tool.setuptools]
 license-files = ["LICENSE"]
@@ -223,7 +229,9 @@ module = [
     "pygtrie.*",
     "pyinstrument",
     "pyparsing",
+    "pytest_benchmark.*",
     "pytest_docker.plugin",
+    "pytest_virtualenv.*",
     "ruamel.*",
     "ruamel.yaml.*",
     "shortuuid",
